Created onClick and onShiftClick properties on DocumentItem and DocumentCards components so you can use them in different flows and have different onClick and onShiftClick behaviors. Before you always need to shift + click to select, which was not intuitive in some situations, such as the document select for extracts.
